fix the issue

+++ controllers/vote.rb
@@ -5,6 +5,11 @@ end
 
 delete '/vote/:id/delete' do 
     user_vote = Vote.find_by(answer_id: params[:id], user_id: current_user.id)
-    user_vote.destroy
-    redirect to("/users/#{current_user.id}")
+    if user_vote 
+        user_vote.destroy
+        redirect to("/users/#{current_user.id}")
+    else 
+        redirect to("/users/#{current_user.id}")
+    end
+
 end
